pyproject.toml
[tool.poetry]
name = "vicon"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "icon_sim"}]

[tool.poetry.dependencies]
python = "^3.11"
black = "^24.4.2"
parameterized = "^0.9.0"
torch = "^2.3.1"
omegaconf = "^2.3.0"
torchvision = "^0.18.1"
numpy = "^1.0"
wandb = "^0.19.9"
hydra-core = "^1.3.2"
matplotlib = "^3.10.1"
pytz = "^2025.2"
h5py = "^3.13.0"
torchdata = "^0.11.0"
absl-py = "^2.2.2"
tabulate = "^0.9.0"
torchinfo = "^1.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
